Complete splitting up probdiffeq
