+++ probdiffeq/_probdiffeq/estimation_strategies.py
@@ -1,0 +1,453 @@
+class MarkovStrategy(Generic[T]):
+    """An interface for estimation strategies in Markovian state-space models.
+
+    Related:
+    [`strategy_filter`](#probdiffeq.probdiffeq.strategy_filter),
+    [`strategy_smoother_fixedpoint`](#probdiffeq.probdiffeq.strategy_smoother_fixedpoint),
+    [`strategy_smoother_fixedinterval`](#probdiffeq.probdiffeq.strategy_smoother_fixedinterval).
+    """
+
+    def __init__(
+        self,
+        ssm: ssm_impl.FactSsmImpl,
+        is_suitable_for_save_at: int,
+        is_suitable_for_save_every_step: int,
+        is_suitable_for_offgrid_marginals: int,
+    ) -> None:
+        self.ssm = ssm
+        self.is_suitable_for_save_at = is_suitable_for_save_at
+        self.is_suitable_for_save_every_step = is_suitable_for_save_every_step
+        self.is_suitable_for_offgrid_marginals = is_suitable_for_offgrid_marginals
+
+    def init_posterior(self, *, u) -> T:
+        """Initialize a posterior distribution."""
+        raise NotImplementedError
+
+    def predict(self, posterior: T, *, transition) -> tuple:
+        """Make a prediction."""
+        raise NotImplementedError
+
+    def apply_updates(self, prediction: T, *, updates) -> tuple:
+        """Apply updates to a prediction."""
+        raise NotImplementedError
+
+    def interpolate(
+        self, *, posterior_t0: T, posterior_t1: T, transition_t0_t, transition_t_t1
+    ) -> tuple[tuple, InterpResult[T]]:
+        """Interpolate between two points."""
+        raise NotImplementedError
+
+    def interpolate_at_t1(self, *, posterior_t1: T) -> tuple[tuple, InterpResult[T]]:
+        """Interpolate at a checkpoint."""
+        raise NotImplementedError
+
+    def finalize(self, *, posterior0: T, posterior: T, output_scale) -> T:
+        """Finalize the posterior before returning a solution."""
+        raise NotImplementedError
+
+
+class strategy_smoother_fixedinterval(MarkovStrategy[MarkovSequence]):
+    """Construct a fixed-interval smoother.
+
+    Use this strategy for fixed steps.
+    For adaptive steps, consider using a fixed-point smoother instead.
+
+
+    Related:
+    [`MarkovStrategy`](#probdiffeq.probdiffeq.MarkovStrategy).
+    """
+
+    def __init__(self, ssm) -> None:
+        super().__init__(
+            ssm=ssm,
+            is_suitable_for_save_at=False,
+            is_suitable_for_save_every_step=True,
+            is_suitable_for_offgrid_marginals=True,
+        )
+
+    def init_posterior(self, *, u):
+        cond = self.ssm.prior.identity(template=u)
+        posterior = MarkovSequence(marginal=u, conditional=cond, reverse=True)
+        return u, posterior
+
+    def predict(self, posterior: MarkovSequence, *, transition) -> tuple:
+        marginals, cond = self.ssm.conditional.revert(
+            posterior.marginal, transition, solve_triu=linalg.solve_triu
+        )
+        posterior = MarkovSequence(
+            marginal=marginals, conditional=cond, reverse=posterior.reverse
+        )
+
+        return marginals, posterior
+
+    def apply_updates(self, prediction, *, updates):
+        posterior = MarkovSequence(
+            updates, prediction.conditional, reverse=prediction.reverse
+        )
+        marginals = updates
+        return marginals, posterior
+
+    def finalize(
+        self, *, posterior0: MarkovSequence, posterior: MarkovSequence, output_scale
+    ):
+        prototype = self.ssm.prior.prototype_output_scale_calibrated(
+            template=posterior0.marginal
+        )
+        assert output_scale.shape == prototype.shape
+        posterior0 = posterior0.rescale_cholesky(output_scale)
+        posterior = posterior.rescale_cholesky(output_scale)
+
+        # Marginalise
+        marginals = posterior.evaluate_marginals(ssm=self.ssm)
+
+        # Prepend the initial condition to the filtering distributions
+        init = tree.tree_array_prepend(posterior0.marginal, posterior.marginal)
+        posterior = MarkovSequence(
+            marginal=init, conditional=posterior.conditional, reverse=posterior.reverse
+        )
+
+        # Extract targets
+        return marginals, posterior
+
+    def interpolate(
+        self,
+        *,
+        posterior_t0: MarkovSequence,
+        posterior_t1: MarkovSequence,
+        transition_t0_t,
+        transition_t_t1,
+    ):
+        """Interpolate between two Markov sequences.
+
+        Here is how a smoother interpolates:
+
+        - Extrapolate from t0 to t, which gives the filtering distribution
+          and the backward transition from t to t0.
+        - Extrapolate from t to t1, which gives another filtering distribution
+          and the backward transition from t1 to t.
+        - Apply the new t1-to-t backward transition to the posterior
+          to compute the interpolation.
+
+        This intermediate result is informed about its "right-hand side" datum.
+        Note how in probdiffeq, all solver subintervals include their right-hand
+        side time-point: in other words, they are (t0, t1].
+
+        Specifically, interpolation is not equal to computing offgrid marginals.
+        Interpolation always assumes that the current subinterval is the right-most
+        subinterval, which is the case during the forward pass.
+        After the simulation, if there are observations > t1,
+        which happens when computing offgrid-marginals, do not use `interpolate()`.
+        """
+        # Extrapolate from t0 to t, and from t to t1.
+
+        _, extrapolated_t = self.predict(
+            posterior=posterior_t0, transition=transition_t0_t
+        )
+        _, extrapolated_t1 = self.predict(
+            posterior=extrapolated_t, transition=transition_t_t1
+        )
+
+        # Marginalise backwards from t1 to t to obtain the interpolated solution.
+        marginal_t1 = posterior_t1.marginal
+        conditional_t1_to_t = extrapolated_t1.conditional
+        rv_at_t = self.ssm.conditional.marginalise(marginal_t1, conditional_t1_to_t)
+        solution_at_t = MarkovSequence(
+            rv_at_t, extrapolated_t.conditional, reverse=extrapolated_t.reverse
+        )
+
+        # The state at t1 gets a new backward model;
+        # (it must remember how to get back to t, not to t0).
+        solution_at_t1 = MarkovSequence(
+            marginal_t1, conditional_t1_to_t, reverse=extrapolated_t.reverse
+        )
+        interp_res = InterpResult(step_from=solution_at_t1, interp_from=solution_at_t)
+
+        # Extract targets
+        marginals = solution_at_t.marginal
+        return (marginals, solution_at_t), interp_res
+
+    def interpolate_at_t1(self, posterior_t1):
+        marginals = posterior_t1.marginal
+
+        interp_res = InterpResult(step_from=posterior_t1, interp_from=posterior_t1)
+        return (marginals, posterior_t1), interp_res
+
+
+class strategy_filter(MarkovStrategy):
+    """Construct a filter.
+
+    Filters work with all timestepping strategies.
+    However, the uncertainties are not informed by the full
+    timeseries, which makes them visually less pleasing.
+    Filter solutions also do not admit computing log-marginal
+    likelihoods or joint sampling from the posterior distribution.
+    For these use-cases, use smoothers instead.
+
+    Related:
+    [`MarkovStrategy`](#probdiffeq.probdiffeq.MarkovStrategy).
+    """
+
+    def __init__(self, ssm) -> None:
+        super().__init__(
+            ssm=ssm,
+            is_suitable_for_save_at=True,
+            is_suitable_for_save_every_step=True,
+            is_suitable_for_offgrid_marginals=True,
+        )
+
+    def init_posterior(self, *, u):
+        return u, u
+
+    def predict(self, posterior: T, *, transition) -> tuple:
+        marginals = self.ssm.conditional.marginalise(posterior, transition)
+        return marginals, marginals
+
+    def apply_updates(self, prediction, *, updates):
+        del prediction
+        marginals = updates
+        return marginals, marginals
+
+    def finalize(self, *, posterior0, posterior, output_scale):
+        expected = self.ssm.prior.prototype_output_scale_calibrated(template=posterior0)
+        assert output_scale.shape == expected.shape
+
+        # No rescaling because no calibration at the initial step
+        posterior0 = posterior0.rescale_cholesky(output_scale)
+
+        # Calibrate
+        posterior = posterior.rescale_cholesky(output_scale)
+
+        # Stack
+        posterior = tree.tree_array_prepend(posterior0, posterior)
+
+        marginals = posterior
+        return marginals, posterior
+
+    def interpolate(
+        self, *, posterior_t0, posterior_t1, transition_t0_t, transition_t_t1
+    ):
+        del transition_t_t1
+        _, interpolated = self.predict(
+            posterior=posterior_t0, transition=transition_t0_t
+        )
+        marginals = interpolated
+        interp_res = InterpResult(step_from=posterior_t1, interp_from=interpolated)
+        return (marginals, interpolated), interp_res
+
+    def interpolate_at_t1(self, *, posterior_t1):
+        marginals = posterior_t1
+        interp_res = InterpResult(step_from=posterior_t1, interp_from=posterior_t1)
+        return (marginals, posterior_t1), interp_res
+
+
+class strategy_smoother_fixedpoint(MarkovStrategy[MarkovSequence]):
+    r"""Construct a fixedpoint-smoother.
+
+    Fixed-point smoothers are the method of choice for adaptive
+    time-stepping in probabilistic differential equation solvers.
+
+    Concretely, we implement the fixedpoint smoother by Krämer (2025a).
+    Applied to probabilistic solvers, this leads to the algorithm
+    by Krämer (2025b). Please consider citing both papers if you use
+    fixed-point smoothers and ODE solvers in your work.
+
+
+    ??? note "BibTex for Krämer (2025a)"
+        ```bibtex
+        @article{kramer2025numerically,
+            title={Numerically Robust Fixed-Point Smoothing Without State Augmentation},
+            author={Kr{\"a}mer, Nicholas},
+            year={2025},
+            journal={Transactions on Machine Learning Research}
+        }
+        ```
+
+    ??? note "BibTex for Krämer (2025b)"
+        ```bibtex
+            @InProceedings{kramer2024adaptive,
+            title     = {Adaptive Probabilistic ODE Solvers Without Adaptive
+            Memory Requirements},
+            author    = {Kr\"{a}mer, Nicholas},
+            booktitle = {Proceedings of the First International Conference
+            on Probabilistic Numerics},
+            pages     = {12--24},
+            year      = {2025},
+            editor    = {Kanagawa, Motonobu and Cockayne, Jon and Gessner,
+            Alexandra and Hennig, Philipp},
+            volume    = {271},
+            series    = {Proceedings of Machine Learning Research},
+            publisher = {PMLR},
+            url       = {https://proceedings.mlr.press/v271/kramer25a.html}
+        }
+        ```
+    Related:
+    [`MarkovStrategy`](#probdiffeq.probdiffeq.MarkovStrategy).
+
+
+    """
+
+    def __init__(self, ssm) -> None:
+        super().__init__(
+            ssm=ssm,
+            is_suitable_for_save_at=True,
+            is_suitable_for_save_every_step=False,
+            is_suitable_for_offgrid_marginals=False,
+        )
+
+    def init_posterior(self, *, u):
+        cond = self.ssm.prior.identity(template=u)
+        posterior = MarkovSequence(u, cond, reverse=True)
+        return u, posterior
+
+    def predict(self, posterior: MarkovSequence, *, transition) -> tuple:
+        rv = posterior.marginal
+        bw0 = posterior.conditional
+        marginals, cond = self.ssm.conditional.revert(
+            rv, transition, solve_triu=linalg.solve_triu
+        )
+        cond = self.ssm.conditional.merge(bw0, cond)
+        predicted = MarkovSequence(marginals, cond, reverse=posterior.reverse)
+
+        return marginals, predicted
+
+    def apply_updates(self, prediction: MarkovSequence, *, updates):
+        posterior = MarkovSequence(
+            updates, prediction.conditional, reverse=prediction.reverse
+        )
+        marginals = updates
+        return marginals, posterior
+
+    def finalize(
+        self, *, posterior0: MarkovSequence, posterior: MarkovSequence, output_scale
+    ):
+        expected = self.ssm.prior.prototype_output_scale_calibrated(
+            template=posterior0.marginal
+        )
+        assert output_scale.shape == expected.shape
+        posterior0 = posterior0.rescale_cholesky(output_scale)
+        posterior = posterior.rescale_cholesky(output_scale)
+
+        # Marginalise
+        marginals = posterior.evaluate_marginals(ssm=self.ssm)
+
+        # Prepend the initial condition to the filtering distributions
+        init = tree.tree_array_prepend(posterior0.marginal, posterior.marginal)
+        posterior = MarkovSequence(
+            marginal=init, conditional=posterior.conditional, reverse=posterior.reverse
+        )
+
+        # Extract targets
+        return marginals, posterior
+
+    def interpolate_at_t1(self, *, posterior_t1: MarkovSequence):
+        cond_identity = self.ssm.prior.identity(template=posterior_t1.marginal)
+        resume_from = MarkovSequence(
+            posterior_t1.marginal,
+            conditional=cond_identity,
+            reverse=posterior_t1.reverse,
+        )
+        interp_res = InterpResult(step_from=resume_from, interp_from=resume_from)
+
+        interpolated = posterior_t1
+        marginals = interpolated.marginal
+        return (marginals, interpolated), interp_res
+
+    def interpolate(
+        self,
+        *,
+        posterior_t0: MarkovSequence,
+        posterior_t1: MarkovSequence,
+        transition_t0_t,
+        transition_t_t1,
+    ):
+        """Interpolate between two Markov sequences.
+
+        Assuming `state_t0` has seen $n$ collocation points,
+        and `state_t1` has seen $n+1$ collocation points,
+        then interpolation at time $t$ is computed as follows:
+
+        1. Extrapolate from $t_0$ to $t$. This yields:
+            - the marginal at $t$ given $n$ observations.
+            - the backward transition from $t$ to $t_0$ given $n$ observations.
+
+        2. Extrapolate from $t$ to $t_1$. This yields:
+            - the marginal at $t_1$ given $n$ observations
+              (in contrast,`state_t1` has seen $n+1$ observations)
+            - the backward transition from $t_1$ to $t$ given $n$ observations.
+
+        3. Apply the backward transition from $t_1$ to $t$
+        to the marginal inside `state_t1`
+        to obtain the marginal at $t$ given $n+1$ observations. Similarly,
+        the interpolated solution inherits all auxiliary info from the $t_1$ state.
+
+        ---------------------------------------------------------------------
+
+        All comments from fixed-interval smoother interpolation apply.
+
+        ---------------------------------------------------------------------
+
+        Difference to standard smoother interpolation:
+
+        In the fixed-point smoother, backward transitions are modified
+        to ensure that future operations remain correct.
+        Denote the location of the fixed-point with $t_f$. Then,
+        the backward transition at $t$ is merged with that at $t_0$.
+        This preserves knowledge of how to move from $t$ to $t_f$.
+
+        Then, `t` becomes the new fixed-point location. To ensure
+        that future operations "find their way back to $t$":
+
+        - Subsequent interpolations do not continue from the raw
+        interpolated value. Instead, they continue from a nearly
+        identical state where the backward transition is replaced
+        by the identity.
+
+        - Subsequent solver steps do not continue from the initial $t_1$
+        state. Instead, they continue from a version whose backward
+        model is replaced with the `t-to-t1` transition.
+
+
+        ---------------------------------------------------------------------
+
+        As a result, each interpolation must return three distinct states:
+
+        1. the interpolated solution,
+        2. the state to continue interpolating from,
+        3. the state to continue solver stepping from.
+
+        These are intentionally different in the fixed-point smoother.
+        """
+        # Note to myself: Don't attempt to remove any of them.
+        # They're all important. You will break the code (again) :).
+
+        # Extrapolate from t0 to t, and from t to t1.
+        # This yields all building blocks.
+        _, extrapolated_t = self.predict(
+            posterior=posterior_t0, transition=transition_t0_t
+        )
+        conditional_id = self.ssm.prior.identity(template=posterior_t0.marginal)
+        previous_new = MarkovSequence(
+            extrapolated_t.marginal, conditional_id, reverse=extrapolated_t.reverse
+        )
+        _, extrapolated_t1 = self.predict(
+            posterior=previous_new, transition=transition_t_t1
+        )
+
+        # Marginalise from t1 to t to obtain the interpolated solution.
+        marginal_t1 = posterior_t1.marginal
+        conditional_t1_to_t = extrapolated_t1.conditional
+        rv_at_t = self.ssm.conditional.marginalise(marginal_t1, conditional_t1_to_t)
+
+        # Return the right combination of marginals and conditionals.
+        interpolated = MarkovSequence(
+            rv_at_t, extrapolated_t.conditional, reverse=extrapolated_t.reverse
+        )
+        step_from = MarkovSequence(
+            posterior_t1.marginal,
+            conditional=conditional_t1_to_t,
+            reverse=posterior_t1.reverse,
+        )
+        interp_res = InterpResult(step_from=step_from, interp_from=previous_new)
+
+        marginals = interpolated.marginal
+        return (marginals, interpolated), interp_res
