+++ probdiffeq/_probdiffeq/utilities.py
@@ -1,0 +1,42 @@
+S = TypeVar(
+    "S", bound=ProbabilisticSolution | MarkovSequence | ssm_impl.AbstractTreeNormal
+)
+"""A type-variable to describe interpolation results."""
+
+
+@tree.register_dataclass
+@structs.dataclass
+class InterpResult(Generic[S]):
+    """A datastructure to store interpolation results.
+
+    To ensure correct adaptive time-stepping, it is important
+    to distinguish step-from variables from interpolate-from variables.
+
+    For some solvers, e.g. fixed-point-smoother-based ones,
+    both stepping and interpolating variables are adjusted during interpolation.
+    """
+
+    step_from: S
+    """The new 'step_from' field.
+
+    At time `max(t, s1.t)`.
+    Use this as the right-most reference state
+    in future interpolations, or continue time-stepping from here.
+    """
+
+    interp_from: S
+    """The new `interp_from` field.
+
+    At time `t`. Use this as the left-most reference state
+    in future interpolations.
+
+    The difference between `interpolated` and `interp_from`
+    is important around checkpoints:
+
+    - `interpolated` belongs to the just-concluded time interval,
+    - `interp_from` belongs to the to-be-started time interval.
+
+    Concretely, this means that for fixed-point smoothers,
+    `interp_from` has a unit backward model whereas `interpolated`
+    remembers how to step back to the previous target location.
+    """
